修改创建项目、spider 的去重查找方法

+++ tool/create_project.py
@@ -26,6 +26,9 @@ class CreateProject:
 
         parent_path = self.path.joinpath(self.project_name.title())
 
+        if parent_path.exists():
+            raise FileExistsError(f'{parent_path} 已存在！')
+
         shutil.copytree(self.path_template, parent_path)
 
     def check_name(self):
